refactoring: use ammPrices instead of pancake and nyanswop
