+++ src/api/price/index.js
@@ -1,11 +1,8 @@
-const { getAmmLpPrices } = require('../../api/stats/pancake/getCakePrices');
+const { getAmmTokensPrices, getAmmLpPrices } = require('../stats/getAmmPrices');
 
-// TODO: Remove imports below
+// TODO: Remove all imports below in favor of getAmmPrices
 const { lpTokenPrices } = require('../../utils/lpTokens');
 const fetchPrice = require('../../utils/fetchPrice');
-const { getNyanswopTokenPrices } = require('../stats/nyanswop/getNyanswopPrice');
-const { getCakeTokensPrices } = require('../stats/pancake/getCakePrices');
-const cakeLpTokens = require('../../data/cakeLpPools.json');
 const thugsLpTokens = require('../../data/thugsLpPools.json');
 const bakeryLpTokens = require('../../data/bakeryLpPools.json');
 const narLpTokens = require('../../data/narLpPools.json');
@@ -46,10 +43,6 @@ async function lpPrices(ctx, lpTokens) {
     console.error(err);
     ctx.status = 500;
   }
-}
-
-async function cakeLpPrices(ctx) {
-  await lpPrices(ctx, cakeLpTokens);
 }
 
 async function thugsLpPrices(ctx) {
@@ -134,7 +127,7 @@ async function bakeryPrices(ctx) {
 
 async function nyanswopPrices(ctx) {
   try {
-    const prices = await getNyanswopTokenPrices();
+    const prices = await getAmmTokensPrices();
     ctx.status = 200;
     ctx.body = prices;
   } catch (err) {
@@ -143,9 +136,9 @@ async function nyanswopPrices(ctx) {
   }
 }
 
-async function pancakePrices(ctx) {
+async function tokenPrices(ctx) {
   try {
-    const prices = await getCakeTokensPrices();
+    const prices = await getAmmTokensPrices();
     ctx.status = 200;
     ctx.body = prices;
   } catch (err) {
@@ -155,7 +148,7 @@ async function pancakePrices(ctx) {
 
 module.exports = {
   lpsPrices,
-  cakeLpPrices,
+  tokenPrices,
   thugsLpPrices,
   bakeryLpPrices,
   narLpPrices,
@@ -167,7 +160,6 @@ module.exports = {
   kebabLpPrices,
   monsterLpPrices,
   nyanswopLpPrices,
-  pancakePrices,
   spongeLpPrices,
   boltLpPrices,
   autoLpPrices,
